add EdgeLayer struct with ingestion implementation
